+++ MEKB_H0_Anlage/Hauptform/Form1.resx
@@ -315,6 +315,13 @@
         vgAAAABJRU5ErkJggg==
 </value>
   </data>
+  <data name="button1.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+    <value>
+        iVBORw0KGgoAAAANSUhEUgAAAB8AAAAPBAMAAADuVgmRAAAABGdBTUEAALGPC/xhBQAAAAZQTFRFgICA
+        QEBAEVco6wAAAAlwSFlzAAAOwQAADsEBuJFr7QAAABVJREFUGNNjIAIIIgOBISWAHzAwAAAT+gows0AL
+        vgAAAABJRU5ErkJggg==
+</value>
+  </data>
   <metadata name="label1.Locked" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
     <value>True</value>
   </metadata>
